Unorganized improvements
i have made strides at improving the layout of the pages and allowing for further expansion of this app to used for any form. I have moved the buttons for the new and edit pages for CIARs and Change Request to dedicated home pages. All that remains at / is the button to enter these new landing pages.